move chipingresseset to a dedicated go module to limit blast radius of extra dependencies
