+++ framework/components/dockercompose/chip_ingress_set/docker-compose.yml
