+++ framework/components/dockercompose/chip_ingress_set/redpanda.go
